Lighters!
TO DO: Add a delay variable to the parameters on the robo-fix helper
proc that handles this, and allow different tools a different delay.
Make lighters take far longer and for less HP restored each go, and
similarly, make each surgical repair once they're past the threshold
take longer and be less effective.