Duct-tape as a ghetto medical tool!
This tweak allows duct-tape to serve as tier-1 first aid and allow it to
perform minor repairs to damaged synthetic limbs.

Of note is that the duct-tape does stop bleeding, but does not disinfect
wounds, this is intentional to keep with its intended use as a last
resort or crafty patch-up job.

My only real concern so far is that duct tape isn't a limited use item
like a cable coil would be, but if that's judged to be of concern I can
make that change and open another request with a limited number of uses.